refactor modules to use domain specific generators
