+++ src/decorators/languageVariantsDecorator.js
@@ -1,4 +1,4 @@
-const normalize = require(`./normalize`);
+const normalize = require(`../normalize`);
 
 /**
  * Add Gatsby relations from item nodes to it other language variants.
@@ -44,27 +44,6 @@ const decorateItemsWithLanguageVariants = (
   }
 };
 
-/**
- * Add Gatsby relations from type nodes to items based on this type.
- * @param {Array} contentItemNodes
- *  Gatsby content item nodes to make a link in
- * @param {Array} contentTypeNodes
- *  Gatsby content type nodes
- */
-const decorateTypeNodesWithItemLinks = (
-  contentItemNodes,
-  contentTypeNodes) => {
-  try {
-    normalize.decorateTypeNodesWithItemLinks(
-      contentItemNodes,
-      contentTypeNodes
-    );
-  } catch (error) {
-    console.error(error);
-  }
-};
-
 module.exports = {
   decorateItemsWithLanguageVariants,
-  decorateTypeNodesWithItemLinks,
 };
